feature improvement: gottcha2 fasta upload removed duplicate entries
For each extraction of the gottcha2 fastas there is an entry for each
level of the taxonomic tree. It only duplicates the sequence and changes
the LEVEL=[genus, species, kingdom, etc] in the header. I have removed
all of these duplicate entries except on the lowest level per read which
is typically strain but can also be species if there does not exist a
strain entry. Typically this is from a entry in the gottcha2 db for a
species without strain level info.