properly handling the human read scrub parameter when it is null
The clumpify workflow runs checks to verify that a provided human read
scrubbing database is a file that exists. However, the previous
implementation did not account for the possibility that the parameter
would be null, which is in fact its default. Instead, the first check
was:

```nextflow

if (file(params.human_read_scrub),exists()) {
    ...
}
```

This commit fixes that, making sure that the parameter is not null
before proceeding to checks that would otherwise break when given a
null.

This commit also supplies human read scrubbing with an empty channel
instead of relying on whether the parameter is specified as a check.
This means that the process will only ever be run if the human scrubbing
file path exists and is a file.