Bug fix: gottcha2 tsv output statistics match old nvd1 version
The issue https://github.com/dhoconno/nvd/issues/8 describes the problem
well along with my comments. In short, we have been using nvd1 with a
old version of gottcha2 that had different output statistics. All of the
information is still there but doesn't all get output or calculated as
the creator of gottcha2 modified what should be output in the full.tsv
file. I renamed a few of the columns to match the old naming convention
we use on labkey and added a few calculations to again calculate
meaningful stats that the creator removed from the full output.