turning off regex mode for seqkit grep's role in the the initial STAT classification step, plus a patch bump release
For very large databases and pattern sets, `seqkit grep` has been taking
unusably long on our HPC cluster to extract reads matching STAT human
virus classifications. On further investigation, it turned out that
we've been running `seqkit grep` in regex mode instead of searching for
the pattern literals directly. Regex mode was not previously used in
this step in NVD1. This commit reverts to that previous use of `seqkit
grep` in hopes of getting performance for human virus read extraction
that matches what we've observed in NVD1.