bug fix: labkey.config file included regarless if --labkey option set
When a user tries to run the pipeline locally and don't have a
labkey.config file located at: ${projectDir}/conf/labkey.config they
will get an error. This is a problem as only select users need to use
the labkey support and primary audiences running this pipeline will
crash immedietly. I made the import of the labkey.config conditional to
the --labkey flag in the run command.