improvement: delete intermediate fasta file in stat step to save space
While running NVD2 on our cluster we have encountered our disk capacity
being exceeded by the nvd portion of the pipeline. This process creates
a .fasta representation of every interleaved .fastq file and only uses
it within the process. There is no reason it needs to be saved in work
and can therefore be deleted.