make select blast rows work with the raw mega/blast results
Select rows for the top N blast hits is being placed right after
megablast and blastn classify. So what we need to do is add a header and
also remove the additional sorting and over methods in polars. There is
no task or sample_id yet in the raw blast output as that is added later
in the PREPARE_BLAST_LABKEY step before upload. After the top X rows are
selected from the blast tasks we also must remove the intermediate
header we added for polars to detect the columns. So we use the
next(reader) function as the first step when using the annotate python
script.