guard list checking with sessionId or uuid ternary to allow users to set uuid
If there is no uuid param from the labkey.config file or specified on
the CLI then it will use workflow.sessionId. If this is the case then if
the pipeline crashes and the cache for the first part is destroyed then
a new sessionId will be computed. If so the user will get a fail at the
first step (guard_list verification) and will either need to delete the
experimentId from the guard list to input a new sessionID (not
recommended) or use a --uuid that matches the already set sessionId of
previous.