annotate blast results speedup
This lookup step to find a taxid's assocaiated lineage takes a long time
and has many redundant searches for the same taxid. I added a cache that
will store the taxid as a key and its lineage as a value and always
check if the taxid has been seen before to prevent duplicate
computation.