implementing a powerful system for idempotently tracking metagenomic hits themselves across runs, samples, etc.
The commit uses a fast sequence-based hashing approach to identify unique hits and takes advantage of the new state management system, adding a couple new tables to the schema in the process. These tables use a disk-efficient encoding to record sequences alongside hits and their occurrences, enabling FASTA and tabular representations to be recovered from their representation in the nvd sqlite database. As with previous commits, this update comes with a command line interface, `nvd hits`, as well as an expansion in `py_nvd` query APIs. Both the CLI and the API it builds on will be expanded in future commits, but for now, they bring the ability to export hits as one of a few tabular formats or as a FASTA for interop with analytical or bioinformatics tooling.